+++ src/typeguard/__init__.py
@@ -22,6 +22,7 @@ from ._importhook import ImportHookManager as ImportHookManager
 from ._importhook import TypeguardFinder as TypeguardFinder
 from ._importhook import install_import_hook as install_import_hook
 from ._memo import TypeCheckMemo as TypeCheckMemo
+from ._utils import Unset as Unset
 
 # Re-export imports so they look like they live directly in this package
 for value in list(locals().values()):