+++ src/typeguard/_functions.py
@@ -8,7 +8,11 @@ from threading import Lock
 from typing import Any, Callable, NoReturn, TypeVar, overload
 
 from ._checkers import BINARY_MAGIC_METHODS, check_type_internal
-from ._config import global_config
+from ._config import (
+    CollectionCheckStrategy,
+    ForwardRefPolicy,
+    TypeCheckConfiguration,
+)
 from ._exceptions import TypeCheckError, TypeCheckWarning
 from ._memo import TypeCheckMemo
 from ._utils import get_stacklevel, qualified_name
@@ -26,16 +30,41 @@ type_checks_suppress_lock = Lock()
 
 
 @overload
-def check_type(value: object, expected_type: type[T]) -> T:
+def check_type(
+    value: object,
+    expected_type: type[T],
+    *,
+    forward_ref_policy: ForwardRefPolicy = ...,
+    typecheck_fail_callback: TypeCheckFailCallback | None = ...,
+    collection_check_strategy: CollectionCheckStrategy = ...,
+) -> T:
     ...
 
 
 @overload
-def check_type(value: object, expected_type: Any) -> Any:
+def check_type(
+    value: object,
+    expected_type: Any,
+    *,
+    forward_ref_policy: ForwardRefPolicy = ...,
+    typecheck_fail_callback: TypeCheckFailCallback | None = ...,
+    collection_check_strategy: CollectionCheckStrategy = ...,
+) -> Any:
     ...
 
 
-def check_type(value: object, expected_type: Any) -> Any:
+def check_type(
+    value: object,
+    expected_type: Any,
+    *,
+    forward_ref_policy: ForwardRefPolicy = TypeCheckConfiguration().forward_ref_policy,
+    typecheck_fail_callback: (TypeCheckFailCallback | None) = (
+        TypeCheckConfiguration().typecheck_fail_callback
+    ),
+    collection_check_strategy: CollectionCheckStrategy = (
+        TypeCheckConfiguration().collection_check_strategy
+    ),
+) -> Any:
     """
     Ensure that ``value`` matches ``expected_type``.
 
@@ -49,23 +78,40 @@ def check_type(value: object, expected_type: Any) -> Any:
     * Forms a :class:`~.TypeCheckMemo` from the current stack frame
     * Calls the configured type check fail callback if the check fails
 
+    Note that this function is independent of the globally shared configuration in
+    :data:`typeguard.config`. This means that usage within libraries is safe from being
+    affected configuration changes made by other libraries or by the integrating
+    application. Instead, configuration options have the same default values as their
+    corresponding fields in :class:`TypeCheckConfiguration`.
+
     :param value: value to be checked against ``expected_type``
     :param expected_type: a class or generic type instance
+    :param forward_ref_policy: see :attr:`TypeCheckConfiguration.forward_ref_policy`
+    :param typecheck_fail_callback:
+        see :attr`TypeCheckConfiguration.typecheck_fail_callback`
+    :param collection_check_strategy:
+        see :attr:`TypeCheckConfiguration.collection_check_strategy`
     :return: ``value``, unmodified
     :raises TypeCheckError: if there is a type mismatch
 
     """
+    config = TypeCheckConfiguration(
+        forward_ref_policy=forward_ref_policy,
+        typecheck_fail_callback=typecheck_fail_callback,
+        collection_check_strategy=collection_check_strategy,
+    )
+
     if type_checks_suppressed or expected_type is Any:
         return
 
     frame = sys._getframe(1)
-    memo = TypeCheckMemo(frame.f_globals, frame.f_locals)
+    memo = TypeCheckMemo(frame.f_globals, frame.f_locals, config=config)
     try:
         check_type_internal(value, expected_type, memo)
     except TypeCheckError as exc:
         exc.append_path_element(qualified_name(value, add_class_prefix=True))
-        if global_config.typecheck_fail_callback:
-            global_config.typecheck_fail_callback(exc, memo)
+        if config.typecheck_fail_callback:
+            config.typecheck_fail_callback(exc, memo)
         else:
             raise
 
@@ -85,18 +131,18 @@ def check_argument_types(
             exc = TypeCheckError(
                 f"{func_name}() was declared never to be called but it was"
             )
-            if global_config.typecheck_fail_callback:
-                global_config.typecheck_fail_callback(exc, memo)
+            if memo.config.typecheck_fail_callback:
+                memo.config.typecheck_fail_callback(exc, memo)
             else:
                 raise exc
 
         try:
-            check_type_internal(value, annotation, memo=memo)
+            check_type_internal(value, annotation, memo)
         except TypeCheckError as exc:
             qualname = qualified_name(value, add_class_prefix=True)
             exc.append_path_element(f'argument "{argname}" ({qualname})')
-            if global_config.typecheck_fail_callback:
-                global_config.typecheck_fail_callback(exc, memo)
+            if memo.config.typecheck_fail_callback:
+                memo.config.typecheck_fail_callback(exc, memo)
             else:
                 raise
 
@@ -104,15 +150,18 @@ def check_argument_types(
 
 
 def check_return_type(
-    func_name: str, retval: T, annotation: Any, memo: TypeCheckMemo
+    func_name: str,
+    retval: T,
+    annotation: Any,
+    memo: TypeCheckMemo,
 ) -> T:
     if type_checks_suppressed:
         return retval
 
     if annotation is NoReturn or annotation is Never:
         exc = TypeCheckError(f"{func_name}() was declared never to return but it did")
-        if global_config.typecheck_fail_callback:
-            global_config.typecheck_fail_callback(exc, memo)
+        if memo.config.typecheck_fail_callback:
+            memo.config.typecheck_fail_callback(exc, memo)
         else:
             raise exc
 
@@ -128,8 +177,8 @@ def check_return_type(
 
         qualname = qualified_name(retval, add_class_prefix=True)
         exc.append_path_element(f"the return value ({qualname})")
-        if global_config.typecheck_fail_callback:
-            global_config.typecheck_fail_callback(exc, memo)
+        if memo.config.typecheck_fail_callback:
+            memo.config.typecheck_fail_callback(exc, memo)
         else:
             raise
 
@@ -137,7 +186,10 @@ def check_return_type(
 
 
 def check_send_type(
-    func_name: str, sendval: T, annotation: Any, memo: TypeCheckMemo
+    func_name: str,
+    sendval: T,
+    annotation: Any,
+    memo: TypeCheckMemo,
 ) -> T:
     if type_checks_suppressed:
         return sendval
@@ -146,8 +198,8 @@ def check_send_type(
         exc = TypeCheckError(
             f"{func_name}() was declared never to be sent a value to but it was"
         )
-        if global_config.typecheck_fail_callback:
-            global_config.typecheck_fail_callback(exc, memo)
+        if memo.config.typecheck_fail_callback:
+            memo.config.typecheck_fail_callback(exc, memo)
         else:
             raise exc
 
@@ -156,8 +208,8 @@ def check_send_type(
     except TypeCheckError as exc:
         qualname = qualified_name(sendval, add_class_prefix=True)
         exc.append_path_element(f"the value sent to generator ({qualname})")
-        if global_config.typecheck_fail_callback:
-            global_config.typecheck_fail_callback(exc, memo)
+        if memo.config.typecheck_fail_callback:
+            memo.config.typecheck_fail_callback(exc, memo)
         else:
             raise
 
@@ -165,15 +217,18 @@ def check_send_type(
 
 
 def check_yield_type(
-    func_name: str, yieldval: T, annotation: Any, memo: TypeCheckMemo
+    func_name: str,
+    yieldval: T,
+    annotation: Any,
+    memo: TypeCheckMemo,
 ) -> T:
     if type_checks_suppressed:
         return yieldval
 
     if annotation is NoReturn or annotation is Never:
         exc = TypeCheckError(f"{func_name}() was declared never to yield but it did")
-        if global_config.typecheck_fail_callback:
-            global_config.typecheck_fail_callback(exc, memo)
+        if memo.config.typecheck_fail_callback:
+            memo.config.typecheck_fail_callback(exc, memo)
         else:
             raise exc
 
@@ -182,8 +237,8 @@ def check_yield_type(
     except TypeCheckError as exc:
         qualname = qualified_name(yieldval, add_class_prefix=True)
         exc.append_path_element(f"the yielded value ({qualname})")
-        if global_config.typecheck_fail_callback:
-            global_config.typecheck_fail_callback(exc, memo)
+        if memo.config.typecheck_fail_callback:
+            memo.config.typecheck_fail_callback(exc, memo)
         else:
             raise
 
@@ -201,8 +256,8 @@ def check_variable_assignment(
     except TypeCheckError as exc:
         qualname = qualified_name(value, add_class_prefix=True)
         exc.append_path_element(f"value assigned to {varname} ({qualname})")
-        if global_config.typecheck_fail_callback:
-            global_config.typecheck_fail_callback(exc, memo)
+        if memo.config.typecheck_fail_callback:
+            memo.config.typecheck_fail_callback(exc, memo)
         else:
             raise
 
@@ -238,8 +293,8 @@ def check_multi_variable_assignment(
             except TypeCheckError as exc:
                 qualname = qualified_name(obj, add_class_prefix=True)
                 exc.append_path_element(f"value assigned to {varname} ({qualname})")
-                if global_config.typecheck_fail_callback:
-                    global_config.typecheck_fail_callback(exc, memo)
+                if memo.config.typecheck_fail_callback:
+                    memo.config.typecheck_fail_callback(exc, memo)
                 else:
                     raise